deployment/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
Pillow
scipy